Remove QueryManager interface and rename SqlQueryManager
This interface had only a single implementation so it wasn't justified to keep around
